+++ srcs/input_to_tokens/split_to_tokens_utils.c
@@ -1,15 +1,5 @@
 #include "../../includes/mini_shell.h"
 
-void	init_split_info(t_split *split_info)
-{
-	split_info->i = 0;
-	split_info->array_index = 0;
-	split_info->len = 0;
-	split_info->brk_flg = 1;
-	split_info->k = 0;
-	split_info->j = 0;
-}
-
 /*
     ! why this function ?  
     To count the number of arrays needed to store the splitted input.
@@ -25,57 +15,41 @@ void	init_split_info(t_split *split_info)
     * if its neither single or double qts, skip the spaces and increment length
 */
 
-// static void	get_arr_len2(int *i, char *str, int *len, int trig)
-// {
-// 	if (trig == 2)
-// 	{
-// 		(*i)++;
-// 		if (str[*i] == 0)
-// 			(*len)++;
-// 	}
-// 	else if (trig == 1)
-// 	{
-// 		(*i)++;
-// 		(*len)++;
-// 	}
-// }
-
-// int	get_arr_len(char *str)
-// {
-// 	int	len;
-// 	int	i;
-
-// 	i = 0;
-// 	len = 0;
-// 	while (str[i])
-// 	{
-// 		if (str[i] == ' ')
-// 			get_arr_len2(&i, str, &len, 1);
-// 		else if (str[i] != ' ' && str[i])
-// 		{
-// 			if (str[i] == '\"')
-// 			{
-// 				if (!skip_and_increment_len_dbl_qts (str, &i, &len))
-// 					break ;
-// 			}
-// 			else if (str[i] == '\'')
-// 			{
-// 				if (!skip_and_increment_len_sngl_qts (str, &i, &len))
-// 					break ;
-// 			}
-// 			else
-// 				get_arr_len2(&i, str, &len, 2);
-// 		}
-// 	}
-// 	return (len);
-// }
-
 void	inrement_i_len(t_split *split_info)
 {
 	split_info->i++;
 	split_info->len++;
 }
 
+static void	create_string_in_between_qtes(char *str,
+			t_split *split_info)
+{
+	if (str[split_info->i] == '\'')
+	{
+		inrement_i_len(split_info);
+		while (str[split_info->i] && str[split_info->i] != '\'')
+			inrement_i_len(split_info);
+		if (str[split_info->i] == '\'')
+			inrement_i_len(split_info);
+	}
+	else if (str[split_info->i] == '\"')
+	{
+		inrement_i_len(split_info);
+		while (str[split_info->i] && str[split_info->i] != '\"')
+			inrement_i_len(split_info);
+		if (str[split_info->i] == '\"')
+			inrement_i_len(split_info);
+	}
+	else
+		inrement_i_len(split_info);
+}
+
+static void	created_string(char *str, t_split *split_info)
+{
+	split_info->arr[split_info->array_index++] = ft_substr(str,
+			split_info->i - split_info->len, split_info->len);
+	split_info->len = 0;
+}
 /*
     ! why this function ?  
     check if there is closing quote in the string and create new string of characters in between
@@ -92,6 +66,7 @@ void	inrement_i_len(t_split *split_info)
     returns 0 if there is no closing qutoe
     returns 1 if there is a closing quote
 */
+
 int	create_string_in_between_dbl_quotes(char *str, t_split *split_info)
 {
 	inrement_i_len(split_info);
@@ -104,37 +79,14 @@ int	create_string_in_between_dbl_quotes(char *str, t_split *split_info)
 		while (str[split_info->i] && (str[split_info->i] != ' ') && (str[
 					split_info->i] != '<') && (str[split_info->i] != '>')
 			&& (str[split_info->i] != '|'))
-		{
-			if (str[split_info->i] == '\'')
-			{
-				inrement_i_len(split_info);
-				while (str[split_info->i] && str[split_info->i] != '\'')
-					inrement_i_len(split_info);
-				if (str[split_info->i] == '\'')
-					inrement_i_len(split_info);
-			}
-			else if (str[split_info->i] == '\"')
-			{
-				inrement_i_len(split_info);
-				while (str[split_info->i] && str[split_info->i] != '\"')
-					inrement_i_len(split_info);
-				if (str[split_info->i] == '\"')
-					inrement_i_len(split_info);
-			}
-			else
-				inrement_i_len(split_info);
-		}
-		split_info->arr[split_info->array_index++] = ft_substr(str,
-				split_info->i - split_info->len, split_info->len);
-		split_info->len = 0;
+			create_string_in_between_qtes(str, split_info);
+		created_string(str, split_info);
 	}
 	else
 	{
 		while (str[split_info->i] && str[split_info->i] != '\"')
 			inrement_i_len(split_info);
-		split_info->arr[split_info->array_index++] = ft_substr(str,
-				split_info->i - split_info->len, split_info->len);
-		split_info->len = 0;
+		created_string(str, split_info);
 		split_info->brk_flg = 0;
 		return (0);
 	}
@@ -157,6 +109,7 @@ int	create_string_in_between_dbl_quotes(char *str, t_split *split_info)
     returns 0 if there is no closing qutoe
     returns 1 if there is a closing quote
 */
+
 int	create_string_in_between_sngl_quotes(char *str, t_split *split_info)
 {
 	inrement_i_len(split_info);
@@ -169,37 +122,14 @@ int	create_string_in_between_sngl_quotes(char *str, t_split *split_info)
 		while (str[split_info->i] && (str[split_info->i] != ' ')
 			&& (str[split_info->i] != '<') && (str[split_info->i] != '>')
 			&& (str[split_info->i] != '|'))
-		{
-			if (str[split_info->i] == '\'')
-			{
-				inrement_i_len(split_info);
-				while (str[split_info->i] && str[split_info->i] != '\'')
-					inrement_i_len(split_info);
-				if (str[split_info->i] == '\'')
-					inrement_i_len(split_info);
-			}
-			else if (str[split_info->i] == '\"')
-			{
-				inrement_i_len(split_info);
-				while (str[split_info->i] && str[split_info->i] != '\"')
-					inrement_i_len(split_info);
-				if (str[split_info->i] == '\"')
-					inrement_i_len(split_info);
-			}
-			else
-				inrement_i_len(split_info);
-		}
-		split_info->arr[split_info->array_index++] = ft_substr(str,
-				split_info->i - split_info->len, split_info->len);
-		split_info->len = 0;
+			create_string_in_between_qtes(str, split_info);
+		created_string(str, split_info);
 	}
 	else
 	{
 		while (str[split_info->i] && str[split_info->i] != '\'')
 			inrement_i_len(split_info);
-		split_info->arr[split_info->array_index++] = ft_substr(str,
-				split_info->i - split_info->len, split_info->len);
-		split_info->len = 0;
+		created_string(str, split_info);
 		split_info->brk_flg = 0;
 		return (0);
 	}
